requirements.txt
flask==2.3.2
python-dotenv==1.0.1
langchain==0.0.318
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
tqdm==4.65.0
requests==2.31.0